POC for target to runnable mojo.
Result of this mojo should be the same as target to mirror, but final
bits are runnable not p2 repository and generated target would be
referencing this runnable folders as install.

one more case to look at is to have update goal for it so there would be
no need to rebuild whole runnable after change, but just update it.